==> data_analysis_process/4.conclusions_quiz.ipynb <==
# Drawing Conclusions Quiz
Use the space below to explore `store_data.csv` to answer the quiz questions below.

# imports and load data
import pandas as pd
df = pd.read_csv("store_data.csv")

# explore data
df.info()


df['week']= pd.to_datetime(df['week']) 
df.info()

# total sales for the last month
df.iloc[196:, 1:].sum()

# average sales
df.mean()

# sales on march 13, 2016
df[df["week"]=="2016-03-13"]

# worst week for store C
df[df["storeC"]==df["storeC"].min()]

# total sales during most recent 3 month period
last_three_months = df[df['week'] >= '2017-12-01']
# last_three_months.iloc[:, 1:].sum()  # exclude sum of week column
last_three_months.loc[:,"storeA":].sum()

last_three_months = df[df['week'] >= '2017-12-01']

last_three_months.iloc[:, 1:].sum()  # exclude sum of week column

last_three_months

last_three_months.loc[196:,:].sum()
